Make it work with primitive types
