+++ test_programs/compile_success_empty/primitive_trait_method_call_multiple_candidates/Nargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "primitive_trait_method_call_multiple_candidates"
+type = "bin"
+authors = [""]
+
+[dependencies]
