+++ test_programs/compile_success_empty/primitive_trait_method_call_multiple_candidates/src/main.nr
@@ -1,0 +1,7 @@
+fn main() {
+    let x: u64 = 0;
+    let f = Field::from;
+    let _ = f(x);
+
+    let _ = str::from([1, 2, 3]);
+}
